requirements.txt
streamlit
autogen
requests
python-docx 